+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Add static code diagnostics `always-remove-listener`, `avoid-wrapping-in-padding` and `avoid-unnecessary-setstate`.
 * Remove deprecated `Lines of Executable Code` metric, use `Source lines of Code` instead.
 * Add static code diagnostic `avoid-wrapping-in-padding`.
 
@@ -11,7 +12,7 @@
 
 ## 3.2.1
 
-* Remove unnecessary scan by `Lines of Executable Code`
+* Remove unnecessary scan by `Lines of Executable Code`.
 
 ## 3.2.0